Add inset label input and pill input
